Add support. for Cheering Fans and High Kick Move dice roll dialog creation into its own class

+++ modules/game-model/src/commonMain/kotlin/dk/ilios/jervis/model/Location.kt
@@ -78,11 +78,11 @@ data class FieldCoordinate(val x: Int, val y: Int) : Location {
     }
 
     override fun isOnField(rules: Rules): Boolean {
-        return (x >= 0 && x < rules.fieldWidth.toInt() && y >= 0 && y < rules.fieldHeight.toInt())
+        return (x >= 0 && x < rules.fieldWidth && y >= 0 && y < rules.fieldHeight)
     }
 
     override fun isOutOfBounds(rules: Rules): Boolean {
-        return x < 0 || x >= rules.fieldWidth.toInt() || y < 0 || y >= rules.fieldHeight.toInt()
+        return x < 0 || x >= rules.fieldWidth || y < 0 || y >= rules.fieldHeight
     }
 
     override fun getCornerLocation(rules: Rules): CornerThrowInPosition? {
